move to same package structure as Spark
